Implement asset server (#9)
# What is implemented
* Implemented an asset server over HTTP for server admins to add custom
content.
* Assets are placed in `./config/custom_assets`. At startup, the server
generates zlib-compressed versions of the files in `./assets_cache`. The
manifest and manifest CRC files are also automatically generated in the
cache folder.
* Non-existent files return a 404 response. Improper requests (such as a
bad CRC query parameter) return a 400 response.
* If a CRC is provided in the query parameter, the CRC is checked
against the CRC of the local file and a 404 response is returned when
they do not match. If no query parameter is provided, then the file is
always returned.
* Multiple manifest files are supported. (See `Assets_manifest.txt` and
`assets/assetpack000_manifest.txt` in the CWA client.) Each file can
belong to only one manifest; the deepest applicable manifest is always
chosen. For example, with the config in this PR, files inside
`./assets/` will always belong to `assets/assetpack000_manifest.txt`
because `./assets/` is deeper than `./`. Files in `./test/` would belong
to `Assets_manifest.txt` because there's no manifest defined for
`./test/`.
* The Rust module is called `http`, not something more specific to
assets, because we'll also use this HTTP server to forward auth requests
to the auth server.

## Testing
Added test assets with various levels of nesting and compared responses
to sample responses from 2014.

# What is not yet implemented
* TLS/HTTPS
* Asset server inside the oxide-client proxy that can read local files
* Configurable port for the HTTP server