Fix asset server CRC check (#16)
The CRC check in the asset server currently uses the CRC of the
compressed file plus magic number, but it should use the CRC of the
uncompressed file. I've corrected the CRC calculation and made the CRCs
cached in memory, rather than having to recompute the CRC with every
request.