Allow requests to decompressed assets (#17)
Enable requests for decompressed assets to the asset server. The game
will sometimes request assets that are not compressed (namely the
manifest.crc), so this ensures the asset server can respond in either
format. It seems that the magic number `0xa1b2c3d4` and data size are
likely included only for compressed assets; the game doesn't accept a
decompressed `manifest.crc` when these two fields are included at the
beginning of the response.

Assets are still stored in compressed format in the asset cache since
decompressed requests should be rare/only for small files that are fast
to decompress anyway.