Implement indexing for `GuidTable` (#19)
Implements indexing for `GuidTable` to improve performance when
searching for items in the table. For example, this PR changes the
auto-interact logic to only check NPCs that are actually
auto-interactable. The current implementation works fine because we have
few NPCs, but if there are thousands of NPCs (housing), the server will
likely choke.