Index characters by zone GUID (#21)
Index characters by zone GUID in `GuidTable`. This avoids the need to
search all zones for a specific player, but we also don't need to search
all characters to determine which ones are in the desired zone. In some
cases, this also avoids the need to acquire the zone lock; for example,
we don't need to read-lock the zone every time the player moves. I've
also changed zone/NPC GUID layouts because the global character table
requires a unique GUID for all NPCs between zones. The new layout
imposes a maximum of 65,535 fixtures per house, but we can raise this
limit later by introducing more discriminants for fixtures.