Implement lock enforcer throughout codebase (#24)
* Implements the lock enforcer throughout the codebase.
* Fixes an issue where the lock enforcers would not always lock
characters/zones in the same order because read and write locks were
handled separately.
* Removed an unnecessary zone table lock from the points of interest
code.

**I'm mainly seeking review that the lock enforcer is never called
recursively, and that everything still functions well in-game.** I did
manual testing and everything looks OK, but it's good to get a second
opinion. This is a large PR, so I think line-by-line review won't
provide that much value for the time required.