Implement character chunking (#25)
Implement character chunking. This implementation uses a chunk size of
~200 units, the range at which NPCs start to become small. NPCs within a
3x3 chunk range, centered on the player's current chunk, are visible so
that NPCs no closer than 200 units appear/disappear abruptly.


![image](https://github.com/soir20/oxide/assets/71418127/61c85707-91cb-4a9e-8ca8-c2561804a3e6)

Character chunks are used as part of the index for the character
`GuidTable`. This means that any position update to a character needs to
account for removing and re-inserting the character to update the index
if their chunk changes.