Fix session overwrite race condition (#34)
There is currently a race condition when the client sends two session
requests in quick succession. The client will use the CRC seed generated
from its first session request, but the server will use the second CRC
seed generated, leading to `MismatchedHash` deserialization errors that
prevent the client from connecting. This PR uses the first session
generated rather than the new one.

This does not introduce a security vulnerability because if say, someone
was spoofing UDP packets to impersonate a player and sent a session
request, the session reply still gets sent back to the player and not
the attacker. The attacker doesn't receive information about the current
session. If anything, this prevents an attacker from logging the player
out by spoofing packets to overwrite their session.

Note that "session" does not refer to an authenticated session. This is
referring to the protocol session and not the game server session.

There will be a separate PR at some point to implement proper disconnect
handling to clear sessions.