Handle disconnects: part one (#68)
This PR implements the groundwork for handling disconnects, as well as
these scenarios:
* Client sends a disconnect packet
* Client reconnects with a new session, disconnecting the old session
* Two clients log in as the same player, disconnecting the first client
* The server has reached the maximum number of allowable clients

These scenarios should be enough so that you don't have to kill the
server every time you want to reconnect without changing any config.
This implementation also properly broadcasts packets to everyone nearby
to handle multiplayer scenarios. This will need to be expanded when
friends/group members/squad members log out later on.

Login handling has been split into an authentication step and the actual
login step where the player is added to the world. We need to split them
to prevent adding duplicate players to the world. Currently, the
authentication handler naively assumes the player is who they say they
are for testing purposes.

These scenarios still need to be covered:
* Client stopped responding for some configurable amount of time
* Client's protocol does not match the server's protocol and version
* Client version does not match the server's allowed client versions