Don't disconnect when interacting with missing NPC (#74)
We currently disconnect the player when they interact with a
non-existent NPC. However, this introduces a race condition where the
player may interact with an NPC that is then deleted, but the player's
interact packet may be processed after the deletion. This PR makes the
server simply do nothing instead of sending a disconnect.