Add tickable NPC steps per state (#80)
Allow each tickable NPC state to be broken down into several steps that
are executed in order. This is useful to implement randomized
conversations, for example.

Moved the tickable state to `Character` instead of only being part of
`AmbientNpc` as we will likely need to enable it for other character
types.