Fix tickable NPC animation transitions (#83)
Currently, tickable NPCs do not change their animations properly when
multiple looping animations are in a series. The boolean parameter of
the `SetAnimation` packet determines whether to override the NPC's
animation; it enables looping but does not solely control looping. For
example, if you send a `SetAnimation` packet with a looped-by-default
animation, any following `SetAnimation` packets will be ignored if the
override parameter is `false`.

There seems to be a distinction between spawn animations and regular
animations; we should investigate and update the NPC packet fields
accordingly.