Allow defining randomized transitions between tickable procedures (#89)
Removes the `TickableProcedureOrder` in favor of a more flexible system
where you can specify named procedures to execute next. Each list
contains the names of multiple procedures that could be executed next,
one of which will be selected randomly. The first procedure executed is
also selected randomly from a provided list, if any. An empty or
non-existent list means any of the procedures may be selected.