Support NPC tickable procedure synchronization (#90)
Implements tickable procedure synchronization between NPCs. NPCc can
optionally have an identifying `key` so that the `synchronize_with`
field can reference them by name, rather than GUID or index, which are
not necessarily stable.

The config requires that an NPC with the key provided in
`synchronize_with` exist in the config. Internally, however, the
implementation synchronizes to an NPC with the matching GUID. The GUID
is resolved from the key when processing the config, _not_ when actually
performing the synchronization. (Consider the case where we have
multiple instances of the same zone. We want to synchronize with the NPC
in the same zone, which has a unique GUID but not a unique key.)

We can describe the synchronized NPC as the "follower" and the NPC it is
synchronized to as the "leader." Assume both the leader and follower are
present. Every tick, the follower's current tickable procedure will be
**interrupted** and changed to its own procedure with the same name as
the leader's current procedure. If the follower does not have such a
procedure, or if the leader no longer exists, the follower simply acts
as if it was not synchronized until it is interrupted again.

Tickable procedure trackers now know the last time the procedure changed
to determine whether the follower's procedure needs to be reset.
Furthermore, leaders cannot be the follower of another NPC. In other
words, an NPC cannot be synchronized to another NPC that is
synchronized.

## Why can't an NPC be synchronized to another NPC that is synchronized?
The main issue here is ticking the NPCs in the correct order;
non-synchronized NPCs have to be ticked first. We could support this
behavior if we built a directed acyclic graph (DAG) of all the NPCs
based on the synchronization relationships, but this adds complexity and
requires more memory to store the graph when we can just disallow it.
This also naturally prevents cycles where NPCs are synchronized to each
other.

It also seems cleaner to define one NPC that serves as a source of
events to coordinate all the other NPCs, rather than having transitive
synchronization relationships that would require a DAG to visualize.