Support secondary and tertiary GUID table indices (#108)
Support optional secondary and tertiary indices for items in a GUID
table. For multiplayer chat, we want to be able to efficiently retrieve
a player by name or by squad GUID, which we can't do by selecting on a
range of instance GUIDs and chunks in the current index. Hence, we need
to add two new indices: one with names and one with squad GUIDs.

To conserve memory, items in the GUID table are not added to the
secondary or tertiary indices if they do not have those indices. Hence,
`index2()` and `index3()` return `Option` rather than `()` or some other
direct value.

Note that this PR doesn't yet alter the characters or zone table
indices.