Update client-side unlock status after completing a minigame stage (#115)
Currently, we only update a minigame's unlock status server-side when a
player has completed it. This means that after completing a stage for a
minigame whose stages are unlocked in order, the client prompts the
player to "play again" rather than "play next." We need to send the
stage group instance again after completion to properly update the
unlock status client-side.