Implement multiplayer matchmaking (#121)
* Implement multiplayer matchmaking, including error handling when
failing to find enough players or starting a single-player game when
available.
* Every tick, the game checks for timed out matchmaking groups to decide
whether to start the game (if it has enough players, but not the maximum
number of players) or cancel the game (returning all players to the
world).
* Groups currently have status `OpenToAll` or `OpenToFriends` (unused).
We will use `OpenToFriends` once friend invitations are supported.
* Characters are now indexed by their matchmaking group. The matchmaking
groups are ordered by status, stage group GUID, stage GUID, time of
creation, and owner GUID. This allows us to efficiently select groups
for a given stage that are open to all and timed out, for example.
* This ordering requires us to iterate over all stages during a tick, as
opposed to only selecting timed out groups. (We can select a range based
on stage group, stage, and time, but not a range solely based on time.)
However, the number of stages is a fairly small fixed constant, and
searching for stage config currently requires a linear search anyway, so
iterating over all the stages isn't that inefficient.
* Fixes an issue after #115 where the minigame name stopped showing in
the options menu.
* Adds a `update_value_indices` for GUID tables, which neatly removes
and re-inserts an entry after performing some operation. This is useful
because we have a few places where an error would cause a player to be
removed from the table, but never re-inserted. That case is impossible
when using `update_value_indices`.
* Coordinated start (waiting for all players to be ready) and ending
games when too many players leave will be implemented separately.