Only start minigame when the client is ready (#123)
Currently, we send `CreateActiveMinigame` twice so that the minigame
name appears in the client's options menu. This causes the
`RequestStartActiveMinigame` packet to be sent twice. This PR adds a
check so that the game is not started before the client is ready.

Originally, I planned to implement coordinated game start where we send
start packets to all players at the same time. While testing, I
discovered the play button in the minigame stage UI is re-enabled
whenever a player finishes teleporting, regardless of whether a game is
started or not. This means that the button would be re-enabled while
waiting for other players to finish loading. We should instead handle
such checks on a per-game basis.