Fix missing Flash minigame name, broken how-to button in options menu (#126)
#123 caused the Flash minigame name and how-to button in the options
menu to break. By disabling active minigame creation when the player is
unready, this PR caused one fewer start minigame packet to be received
by the server. This caused the server to send only one Flash game
packet, subtly breaking the options menu because of a wonky client
implementation.

This PR streamlines the minigame pipeline by removing extra packets that
we had to work around this issue. Instead, we only use the minimal
packets necessary now that we better understand why the minigame data
wasn't loading in the options menu:

> Re-send the stage group instance to populate the stage data in the
settings menu.
> When we enter the Flash game HUD state, the current minigame group is
cleared.
> This removes the game name from the options menu and breaks the how-to
button.
> To avoid this, we need to send a Flash game packet to transition the
HUD to the
> Flash game state. Then we re-send the stage group instance data.
Finally, we
> have to reload the Flash game (by sending another Flash game packet)
to ensure
> that stage group instance data is loaded in the options menu.