Properly handle the Flash game end round payload (#127)
Currently, we end the game when receiving the `EndRoundNoValidation`
payload. We should instead grant credits for that round but allow the
game to continue (Stunt Gungan, Turret Tactics).

If we do grant credits at the end of each round, we shouldn't grant any
credits at the end of the game (since they've already been granted).
Furthermore, a game's end score will be the _highest_ score of any
individual round, as in the OG. Otherwise, we'd be awarding players high
scores for simply playing longer and not doing well in each round.