Don't disconnect on `ProcessPacketError`, reduce log spam (#133)
Currently, we disconnect players whenever we have a
`ProcessPacketError`. This is a bit awkward because we tend to treat
some errors as OK when we don't want to disconnect the player, and we
also can only disconnect the player who sent the packet. (In contrast,
we can send packets to any other players via `Broadcast`.)

I don't think disconnecting players on these errors provides much value,
so I've disabled the disconnection (except during login). Since this
makes the "unimplemented packet" log statements a bit longer, I've
marked a few op codes we don't use as ignored to reduce log spam. I've
also introduced a `debug!` macro for debug logging that isn't enabled by
default. This macro is used for the "received packet op code" messages
that are also quite spammy.