Properly handle player house list request (#134)
#133 broke the player house list because we now return an error on
unknown housing op codes. Previously, we always returned an `Ok` result
on unknown op codes, so the house instance list would be added to the
broadcast list on every housing packet received. This PR moves the house
list broadcast to the `RequestPlayerHouseList` handler, rather than the
hacky and now-broken solution of sending it on every housing packet.