Add Hologram cycles (#137)
- Live accurate Hologram cycles to the Ryloth & Jedi Temple War Room
Holotables. Each Hologram lasts 10 seconds before downscaling to an
invisibility period of 6 seconds, after which another hologram is
randomly selected.

- The `TickableProcedure` `tick` function now keeps track of whether the
step is on the first tick. `MoveOnRail` and `UpdateTemporaryAppearance`
packets are only sent if it is the first tick of the step.

- Expanded `spawned_state` to an enum introducing SpawnOnFirstStepTick.
Only players who are in the lobby during the steps first tick will see
the NPC.

---------

Co-authored-by: soir20 <71418127+soir20@users.noreply.github.com>