Refactor miscellaneous `if let` statements to use `let else` (#144)
We currently use a number of nested `if let` statements that have become
a bit unwieldly. I'm refactoring many of these to use a `let else`
pattern to reduce the amount of nesting. This PR refactors `if let
statements` in miscellaneous files where there aren't many `if let`s
used. I'll refactor the other statements in separate PRs.