Fix player being teleported when logging out in matchmaking, merge removal code (#157)
When a player logs out, we currently clear their minigame status and
teleport them, assuming they're actively playing a minigame. However,
it's possible for them to still be in matchmaking, and we teleport them
to their previous position anyway. If we save their position in the
future, this means that logging out during matchmaking could
unexpectedly teleport you to a previous location when you log back in.

I've merged the leave active minigame and matchmaking removal code to
reduce duplication and handle both cases, without needing to check if
the player is in matchmaking or in a minigame everywhere else.

I also removed a check in `RequestCancelActiveMinigame` for a requried
stage GUID. The client only ever requests to exit its own current
minigame that was already tracked by the server. This check really only
would guard against a case where a cancel request was sent absurdly
late, even though the player already entered another minigame. I removed
it to simplify the code.