Refactor chunk and matchmaking ticking (#160)
* Fix an issue where a synchronized character could be updated before
the character it depends on was updated, if the other character was in a
different chunk. All unsynchronized characters are now updated before
any synchronized character is updated.
* Split chunk ticking among a configurable number of threads.
* Split matchmaking ticking from chunk ticking. Matchmaking ticking is
fixed to 1 thread because it requires the characters table write handle,
so all work has to be done synchronously anyway.
* The matchmaking tick period is now 1 second instead of using the 500ms
chunk tick period.
* Set a panic hook so that the entire server exits if any single thread
panics. A panic means that there was some unrecoverable error that we
treat as a crash scenario.
* This is a prerequisite to making certain minigames tickable.