Add prerequisites for daily games (#179)
* Enable minigames that don't teleport the player and don't use the end
screen
* Add daily minigame-specific state packets
* Check the last played status of daily minigames so that the
already-played message is shown in the portal
* We don't gate the stage group instance request on the daily status
because this is handled client-side, besides a server-side check after
"play" is already pressed. We could also add a check when handling a
stage instance request, but this is extra complexity when we won't add
any games that would need it.

The only change to the minigames config is the addition of
`daily_reset_utc_offset_seconds: 0`. The rest is indentation to support
the new top-level field.