Clear closest target after interaction (#202)
After #200, client-side auto-interaction doesn't properly clear the
client's closest interaction target. Send a script to manually clear the
target after processing the interaction. This fixes the issue where NPCs
would not be re-interacted with when entering and leaving the
interaction radius, and auto-interaction with other NPCs would be
prevented.

This does not cause interactions to be repeatedly re-triggered when
standing next to the NPC, though it does if you keep moving within the
NPC's radius. The client only checks if the NPC needs to be interacted
with upon movement.

Test NPCs:

```yaml
    clone_guard:
      model_id: 1936
      pos:
        x: 898.0
        y: 172.44919
        z: 1561.6
        w: 1.0
      rot:
        x: -1
        y: 0
        z: 0
        w: 0
      stand_animation_id: 2505
      cursor: 54
      enable_interact_popup: false
      one_shot_action_on_interact:
          one_shot_action_animation_id: 2506
          duration_millis: 1000
      interact_radius: 6
      auto_interact_radius: 5
    clone_guard2:
      model_id: 1936
      pos:
        x: 896.0
        y: 172.44919
        z: 1561.6
        w: 1.0
      rot:
        x: -1
        y: 0
        z: 0
        w: 0
      stand_animation_id: 2505
      cursor: 54
      enable_interact_popup: false
      one_shot_action_on_interact:
          one_shot_action_animation_id: 2506
          duration_millis: 1000
      interact_radius: 6
      auto_interact_radius: 5
```