NPCs radius fix + durations added to NPC animations + a few additional NPCs (#212)
This solves an issue where the `move_to_interact_offset` was set as the
same as the NPC `interact_radius`, which would cause the animations to
sometimes not work when interacting unless a player is moving closer.
This also adds a duration to the NPC interact animations. Lastly, this
commit adds a few additional NPCs. Here are the NPCs that were added:
**Lounge:**
- Interactible Cannisters

**Main Hall:**
- Interactible Crates

**Members Only:**
- Jar Jar Binks
- Two Clones playing cards at a table (also a NPC)