Support distance-based tickable steps (#220)
* Tickable steps now wait for an NPC to reach its destination before
ending.
* The duration field is now `min_duration_millis` to indicate that the
step could last longer, if the NPC does not reach its destination in
that amount of time.
* The server continually estimates and updates the NPC's position. This
implementation can appropriately handle unexpected speed changes before
a tickable step ends, even if that wasn't part of the server's estimate.
* Sending continuous position updates works around the issue where the
client teleports NPCs if the distance between the old and new positions
is too large.
* This also allows for interactions to happen while the NPC is moving,
even if the NPC is far away from its destination.
* Fixes an issue where NPCs crossing a server-side chunk boundary would
be re-added for clients that could already see the NPC.
* Fixes an issue where the NPC could have a different rotation between
the client and server when its rotation at the destination was set to 0,
0, 0.