Fix tickable step floating point comparison (#221)
It is possible, due to floating point error, that `distance_traveled` is
less than `distance_required`. A better comparison is to check whether
the `old_pos` is set to the destination pos--that means we've confirmed
the NPC reached its destination in `tick`.

(`new_pos` is set equal to the destination pos when the NPC may be able
to reach the destination in the next tick. When the NPC has traveled
enough distance to be able to reach its destination, then the `old_pos`
is set to `new_pos` == destination pos. This implementation isn't
affected by the floating point issue because `max_distance_traveled`
should be larger than any floating point error in
`distance_to_new_pos`.)